fix warnings

+++ src/filter.hpp
@@ -180,7 +180,7 @@ std::vector<T> apply_filter_fma3(const std::vector<T>& samples, const Kernel<T>&
     auto sz = samples.size();
     std::vector<T> filtered;
     filtered.resize(sz);
-    if(kernel.sse_size >= ((sizeof(__m128) / sizeof(T)) * 2)) // make sure we get at least 2 SIMD iterations
+    if((size_t)kernel.sse_size >= ((sizeof(__m128) / sizeof(T)) * 2)) // make sure we get at least 2 SIMD iterations
     {
         for(auto i = 0u; i < sz; ++i)
             filtered[i] = weighted_avg_fma3(samples, kernel, i);
